Many fixes, added merge disable
To disable merging for a subclass of TerrainBlock, set inherited bool
mergeEnabled to false in its initialization e.g. Start().

Many fixes and additions, big and small, to merge system.  At this
point, I'm inclined to call it complete but it has not been extensively
tested and will likely need some tweaking and fixing later.